refactor: share tab utilities
