+++ dashboard.html
@@ -35,6 +35,6 @@
         <section id="categories"></section>
     </main>
     <script type="module" src="https://cdn.jsdelivr.net/npm/emoji-picker-element@^1"></script>
-    <script src="dashboard.js"></script>
-</body>
+    <script type="module" src="dashboard.js"></script>
+  </body>
 </html>
