+++ dashboard.js
@@ -1,4 +1,5 @@
 // Manage categories and tabs on dashboard view
+import { getCategoryData } from './utils.js';
 
 document.addEventListener('DOMContentLoaded', () => {
 const newCatInput = document.getElementById('newCategory');
@@ -25,18 +26,6 @@ themeColorInput.addEventListener('input', () => {
 chrome.storage.sync.set({ themeColor: themeColorInput.value });
 });
 
-function getCategoryData(cats, cat) {
-  if (!cats[cat]) {
-    cats[cat] = { tabs: [], color: '', icon: 'folder' };
-  } else if (Array.isArray(cats[cat])) {
-    cats[cat] = { tabs: cats[cat], color: '', icon: 'folder' };
-  } else {
-    cats[cat].tabs = cats[cat].tabs || [];
-    cats[cat].color = cats[cat].color || '';
-    cats[cat].icon = cats[cat].icon || 'folder';
-  }
-  return cats[cat];
-}
 
 // Load currently open tabs
 function loadOpenTabs() {
@@ -261,19 +250,6 @@ chrome.storage.sync.set({ categories: cats, categoryOrder: order }, loadCategori
 });
 }
 
-// Save current tabs
-function saveTabs(cat) {
-chrome.tabs.query({}, tabs => {
-chrome.storage.sync.get({ categories: {}, categoryOrder: [] }, data => {
-const cats = data.categories;
-const order = data.categoryOrder;
-const catData = getCategoryData(cats, cat);
-catData.tabs = tabs.map(t => ({ url: t.url, title: t.title, favIconUrl: t.favIconUrl || '' }));
-if (!order.includes(cat)) order.push(cat);
-chrome.storage.sync.set({ categories: cats, categoryOrder: order }, loadCategories);
-});
-});
-}
 
 // Open all tabs in category
 function openCategory(cat) {
